More robust chats
- Removed session_status from ChatConnectionMapping
- Added some checks to avoid index errors when
  we just want to delete something (if it's not
  there, silently accept it)

+++ discord/channels.py
@@ -86,7 +86,8 @@ def get_discord_channel(channel_id : str):
 async def delete_discord_channel(channel_id : str):
     guild = server.get_guild()
     channel = guild.get_channel(int(channel_id))
-    await channel.delete()
+    if channel is not None:
+        await channel.delete()
 
 ### Common init functions:
